Refactor to move the checks to find if we have entered via browser back from checkout to signup controller
