+++ client/signup/main.jsx
@@ -70,13 +70,9 @@ import { addP2SignupClassName } from './controller';
 import SiteMockups from './site-mockup';
 import {
 	persistSignupDestination,
-	retrieveSignupDestination,
-	clearSignupDestinationCookie,
 	setSignupCompleteSlug,
 	getSignupCompleteSlug,
-	getSignupCompleteFlowName,
 	setSignupCompleteFlowName,
-	wasSignupCheckoutPageUnloaded,
 } from './storageUtils';
 import {
 	canResumeFlow,
@@ -160,17 +156,12 @@ class Signup extends Component {
 			providedDependencies = pick( queryObject, flow.providesDependenciesInQuery );
 		}
 
-		const searchParams = new URLSearchParams( window.location.search );
-		const isAddNewSiteFlow = searchParams.has( 'ref' );
-
-		if ( isAddNewSiteFlow ) {
-			clearSignupDestinationCookie();
-		}
-
 		// Prevent duplicate sites, check pau2Xa-1Io-p2#comment-6759.
-		if ( ! isAddNewSiteFlow && this.isReEnteringSignupViaBrowserBack() ) {
-			this.enableManageSiteFlow = true;
-			providedDependencies = { siteSlug: getSignupCompleteSlug(), isManageSiteFlow: true };
+		if ( this.props.isManageSiteFlow ) {
+			providedDependencies = {
+				siteSlug: getSignupCompleteSlug(),
+				isManageSiteFlow: this.props.isManageSiteFlow,
+			};
 		}
 
 		this.signupFlowController = new SignupFlowController( {
@@ -284,20 +275,6 @@ class Signup extends Component {
 
 	scrollToTop() {
 		setTimeout( () => window.scrollTo( 0, 0 ), 0 );
-	}
-
-	/**
-	 * Checks if the user entered the signup flow via browser back from checkout page,
-	 * and if they did we will show a modified domain step to prevent creating duplicate sites.
-	 * Check pau2Xa-1Io-p2#comment-6759 for more context.
-	 */
-	isReEnteringSignupViaBrowserBack() {
-		const signupDestinationCookieExists = retrieveSignupDestination();
-		const isReEnteringFlow = getSignupCompleteFlowName() === this.props.flowName;
-		const isReEnteringSignupViaBrowserBack =
-			wasSignupCheckoutPageUnloaded() && signupDestinationCookieExists && isReEnteringFlow;
-
-		return isReEnteringSignupViaBrowserBack;
 	}
 
 	completeP2FlowAfterLoggingIn() {
@@ -666,7 +643,7 @@ class Signup extends Component {
 		const shouldRenderLocaleSuggestions = 0 === this.getPositionInFlow() && ! this.props.isLoggedIn;
 
 		let propsForCurrentStep = propsFromConfig;
-		if ( this.enableManageSiteFlow ) {
+		if ( this.props.isManageSiteFlow ) {
 			propsForCurrentStep = {
 				...propsFromConfig,
 				showExampleSuggestions: false,
